+++ README.md
@@ -1,6 +1,6 @@
 # @MVC 구현하기
 
 ### 2단계 체크리스트
-- [ ] ControllerScanner 클래스에서 @Controller가 붙은 클래스를 찾을 수 있다.
+- [x] ControllerScanner 클래스에서 @Controller가 붙은 클래스를 찾을 수 있다.
 - [ ] HandlerMappingRegistry 클래스에서 HandlerMapping을 처리하도록 구현했다.
 - [ ] HandlerAdapterRegistry 클래스에서 HandlerAdapter를 처리하도록 구현했다.
