+++ mvc/src/main/java/nextstep/mvc/controller/annotation/AnnotationHandlerMapping.java
@@ -1,20 +1,15 @@
 package nextstep.mvc.controller.annotation;
 
 import jakarta.servlet.http.HttpServletRequest;
-import java.lang.reflect.Constructor;
-import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 import java.util.stream.Collectors;
 import nextstep.mvc.HandlerMapping;
-import nextstep.web.annotation.Controller;
 import nextstep.web.annotation.RequestMapping;
 import nextstep.web.support.RequestMethod;
-import org.reflections.Reflections;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -22,7 +17,6 @@ public class AnnotationHandlerMapping implements HandlerMapping {
 
     private static final Logger log = LoggerFactory.getLogger(AnnotationHandlerMapping.class);
     private static final Class<RequestMapping> REQUEST_MAPPING_ANNOTATION_CLASS = RequestMapping.class;
-    private static final Class<Controller> CONTROLLER_ANNOTATION_CLASS = Controller.class;
 
     private final Object[] basePackage;
     private final Map<HandlerKey, HandlerExecution> handlerExecutions;
@@ -33,30 +27,13 @@ public class AnnotationHandlerMapping implements HandlerMapping {
     }
 
     public void initialize() {
-        final Reflections reflections = new Reflections(basePackage);
-        final Set<Class<?>> classes = reflections.getTypesAnnotatedWith(CONTROLLER_ANNOTATION_CLASS);
-
-        for (Class<?> clazz : classes) {
-            final Object instance = getInstance(clazz);
-            final List<Method> methods = getRequestMappingMethods(clazz);
-            methods.forEach(method -> putHandlerExecutionByRequestMapping(instance, method));
+        final ControllerScanner controllerScanner = new ControllerScanner(basePackage);
+        final Map<Class<?>, Object> controllers = controllerScanner.getControllers();
+        for (Class<?> aClass : controllers.keySet()) {
+            final List<Method> methods = getRequestMappingMethods(aClass);
+            methods.forEach(method -> putHandlerExecutionByRequestMapping(controllers.get(aClass), method));
         }
         log.info("Initialized AnnotationHandlerMapping!");
-    }
-
-    private Object getInstance(final Class<?> clazz) {
-        try {
-            final Constructor<?> constructor = clazz.getConstructor();
-            return constructor.newInstance();
-        } catch (NoSuchMethodException exception) {
-            log.error(exception.getMessage());
-            throw new IllegalArgumentException("생성자를 가져올 수 없습니다. " + clazz.getName());
-        } catch (InstantiationException
-                | IllegalAccessException
-                | InvocationTargetException exception){
-            log.error(exception.getMessage());
-            throw new IllegalArgumentException("인스턴스화할 수 없습니다. " + clazz.getName());
-        }
     }
 
     private List<Method> getRequestMappingMethods(final Class<?> clazz) {
